Simplifying
I decided to do away with hashing the key and salting the hash that gets used as the key for the encryption. I got to thinking and all an attacker would need to do is use 0-9 and a-z until it matched the numbers of the hash for what was used as the key. It also didn't really make sense to have it. So I decided to just remove it and make it simpler. 3 arguments now instead of 4, 2 for the encryption and 1 for what you want to encrypt.

The way it will work now can be found the the documentation.md file. The enc_key can be as unique as you want it to be now..instead of just numbers and lowercase a-z.